+++ src/themes/defaultTheme.ts
@@ -1,6 +1,4 @@
-import { createTheme } from '@mui/material/styles'
-
-const defaultTheme = {
+export const defaultTheme = {
     components: {
         MuiAutocomplete: {
             styleOverrides: {
@@ -48,16 +46,3 @@ const defaultTheme = {
         selected: "rgb(0, 127, 255)"
     }
 }
-
-const theme = defaultTheme
-
-type CustomTheme = {
-    [Key in keyof typeof theme]: typeof theme[Key]
-}
-
-declare module "@mui/material/styles" {
-    interface Theme extends CustomTheme { }
-    interface ThemeOptions extends CustomTheme { }
-}
-
-export default createTheme(theme)
